+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/beanSystem/BSDuplicateBeanDefinition.kt
@@ -53,7 +53,7 @@ class BSDuplicateBeanDefinition : AbstractBSInspection() {
 
         val currentFileDeclarations = metas
             .flatMap { it.declarations }
-            .map { it.domAnchor }
+            .mapNotNull { it?.domAnchor }
             .filter { it.containingFile == dom.xmlElement?.containingFile }
             .mapNotNull { it.retrieveDomElement() }
         if (currentFileDeclarations.size > 1) {