+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/beanSystem/BSDuplicateEnumDefinition.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -48,7 +48,7 @@ class BSDuplicateEnumDefinition : AbstractBSInspection() {
         val meta = BSMetaModelAccess.getInstance(project).findMetaForDom(dom) ?: return
 
         val currentFileDeclarations = meta.declarations
-            .map { it.domAnchor }
+            .mapNotNull { it?.domAnchor }
             .filter { it.containingFile == dom.xmlElement?.containingFile }
             .mapNotNull { it.retrieveDomElement() }
         if (currentFileDeclarations.size > 1) {